Health Assignment - Covid
Task 1: We want to demonstrate how the cases are rising (NOTE: case type: confirmed) in the map every day with keeping also the history, using longitude on columns and latitude on rows

Task 2: We want to show the confirmed cases by exact date keeping again all the history.

Task 3: We want to show the confirmed cases in the top 10 countries based on the max date (7th of April).

Task 4: Before we create our dashboard, we need to write down the confirmed cases on the entire view in the center of our sheet. Then create your dashboard.