Sales Assignment
Task 1: We need to demonstrate the number of orders per year using an area chart.
Task 2: We must show the sales per segment and category using bar chart as a default and then change the color of the bars so we can easy see the difference
Task 3: We now need to create a new bin(profit), in order to show the profit in bins of 1k. 
Task 4: We need to show the average discount (showing by percentage in the center of the sheet)
Task 5: Now we need to demonstrate the number of orders by region and category using ship mode.